:recycle: Further reorganized and improved structure of feedback component
